move component to admin folder
